+++ wezterm-gui/src/main.rs
@@ -942,7 +942,7 @@ fn run() -> anyhow::Result<()> {
         opts.config_file.as_ref(),
         &opts.config_override,
         opts.skip_config,
-    );
+    )?;
     let config = config::configuration();
 
     let sub = match opts.cmd.as_ref().cloned() {